Showing alert before delete

+++ src/pages/Users/Users.jsx
@@ -1,21 +1,40 @@
 import { useState } from "react";
 import { useLoaderData } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const Users = () => {
   const loadedUsers = useLoaderData();
   const [users, setUsers] = useState(loadedUsers);
 
   const handleDelete = (id) => {
-    fetch(`http://localhost:3000/users/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          const remaining = users.filter((user) => user._id !== id);
-          setUsers(remaining);
-        }
-      });
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:3000/users/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              const remaining = users.filter((user) => user._id !== id);
+              setUsers(remaining);
+            }
+          });
+
+        Swal.fire({
+          title: "Deleted!",
+          text: "The User has been deleted.",
+          icon: "success",
+        });
+      }
+    });
   };
 
   return (
